+++ CHANGELOG.md
@@ -13,6 +13,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
   - QueryRequest.get/setNumberOfOperations()
   - QueryRequest.get/setOperationNumber()
 - Added new cloud region codes: hsg, abl, dfw, pbv, nbq, ibg, pcz, mez, den, kal
+- Added rowMetadata support, new API for Get/Put/Delete request and result
+  get/set RomMetadata.
+- Added row creation time support, new API Get/Delete Result getCreationTime(), 
+  Put/WriteMultiple Result getExistingCreationTime().
 
 ## [5.4.17] 2025-03-03